Configured ADC by interrupt
The ADC is being controled by the timer A (Timer 5) and it is using
the sequencer no. 1. ADC settings are made in "ADCConfig" function
in skybot.c.

The ADC get the data from PORT E pin no. 3, which is configured in
"ADCConfig" as well

By: Guillermo Gil Rodríguez (guillegilrod@gmail.com)