Add Whisker via interrupt
Whisker has been added using PORT_F with PIN_0, the same as the sw2
on tiva board.

There is also a timer to deal with the bounce of the whisker button.
This timer triggers an interrupt 25ms after a whisker button is pushed.

By: Guillermo Gil Rodríguez (guillegilrod@gmail.com)